fix: 作业执行详情-步骤-主机信息需要显示主机快照数据，而不是当前的数据 #2494
